added docstrings and homogenised function parameters
Now takes string for every argument.  Maybe this is ultimately bad but
for the mean time makes most sense to meas it sort of reflects the GMAD
syntax.  Also added doc strings for method, removed a now non-used
method and removed checks on processes.  There are so many processes and
so decentralised in definition that better just to trust user.