Reworked bias syntax in Madx2Tfs conversion
Before:  Biases were defined and attached to all components in one
step.  This was annoying because you didn't always want to attach to all
of them, but perhaps to only a subset of the components in the lattice.

Now: Biases are defined and attached in separate steps.  Biases are
defined with the keyword argument "biases", and attached to all elements
either with "userdict" (for individual elements) or "allelementdict" to
attach any parameter to all components.  This is more general than the
previous method, and allows you to assign arbitrary parameters and their
values to all components.