Fix wrong aper for madx2gmad w/ a TFS aperdict & sliced tfs beamline
If the beamline TFS is sliced by some means, then the S column will be
recalculated.  An element which was previously at, say, 10km, might
now be at 2m.  So it would look up in the (full) aperture TFS at
S=2m, and this would be wildly off, as the aperture TFS has not been
sliced, leading to the aperture being incorrectly set.

Could perhaps instead also apply the slicing to the aperture instance
as well but this is a more agnostic solution, I think.