colls which are converted to drifts now have correct aper info.
this occurs for example in the LHC model where you will see TAS and
TASB referred to as "RCOLLIMATOR", but really they are not collimators
in any reasonable sense.  they are converted to drifts, as expected,
but without any aperture information, which in all likelihood is
present in the aper model.