Refactored (moved) AddSingleElement out of MadxTfs2Gmad
This is better because it is now much more explicit what exactly
AddSingleElement actually requires (i.e. its parameters).  Before it
was the scope of the whole MadxTfs2Gmad function.  This would
probably be OK if the enclosing function was small, but it is not, so
it is hard to see at a glance what MadxTfs2Gmad actually uses.
It should be more readable / easy to reason about now.