Classes for Builder elements
Breaks no interfaces, fully backwards compatible.  It is better to
leverage python's type system than to rely on a "category" attribute as
this is vastly more extensible.  I need this as I want to be able to
easily split elements for aperture splitting.  The only alternative to
this is a gigantic if/elif.