Restructuring of the Field classes
We added two functions to catch the loop order of a given field map
and to flip a field map if one wants to change the loop order.
Now, a Field doesn't have flip and loop order as class variables anymore.
Internally, the data is always converted to follow the standard loop order
which is xyzt.
When writing out a field map, one can specify the loop order (one of xyzt or tzyx)
to write it out in the required way.